server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "eth-account>=0.13.4",
    "fastapi>=0.115.6",
    "numpy>=2.2.0",
    "pandas>=2.2.3",
    "pyjwt>=2.10.1",
    "torch>=2.5.1",
    "uvicorn>=0.32.1",
    "web3>=7.6.0",
]
